refactor: [#292] 기존의 지역부분을 관리하던 useEffect, useState 방식을 useQuery를 이용하여 캐싱하도록 코드를 리팩토링
